feat: Use only first geom for the moment

+++ src/training.py
@@ -38,8 +38,8 @@ GAMMA = training_params["gamma"]
 
 
 # ========== Input Data ==========
-adj_matrices = np.load(input_data_paths["adj_matrices"], allow_pickle=True).astype(np.int64)
-features = np.load(input_data_paths["features"], allow_pickle=True).astype(np.float32)
+adj_matrices = np.load(input_data_paths["adj_matrices"], allow_pickle=True)
+features = np.load(input_data_paths["features"], allow_pickle=True)
 training_data_size = len(adj_matrices)
 
 np.set_printoptions(threshold=np.inf, linewidth=200)
